testmatrix+test/smoke: moves/renames
