+++ test/smoke/testmain_test.go
@@ -13,21 +13,21 @@ import (
 // sup is the global matrix supervisor, used to collate all test results.
 var sup *testmatrix.Supervisor
 
-// matrixRunner is a wrapper around Runner allowing fully baked fixtures to be
+// runner is a wrapper around Runner allowing fully baked fixtures to be
 // passed directly to tests, rather than the test having to unwrap scenarios
 // themselves.
-type matrixRunner struct{ r *testmatrix.Runner }
+type runner struct{ *testmatrix.Runner }
 
 // Run is analogous to Runner.Run, but accepts a func in terms of strongly typed
 // fixture rather than having to manually unwrap scenarios.
-func (m *matrixRunner) Run(name string, test func(*testing.T, *testFixture)) {
-	m.r.Run(name, func(t *testing.T, c testmatrix.Context) { test(t, c.F.(*testFixture)) })
+func (r *runner) Run(name string, test func(*testing.T, *testFixture)) {
+	r.Runner.Run(name, func(t *testing.T, c testmatrix.Context) { test(t, c.F.(*testFixture)) })
 }
 
-// newMatrixRunner should be called once at the start of every top-level package
+// newRunner should be called once at the start of every top-level package
 // test to produce that test's matrixRunner.
-func newMatrixRunner(t *testing.T, m testmatrix.Matrix) matrixRunner {
-	return matrixRunner{r: sup.NewRunner(t, m)}
+func newRunner(t *testing.T, m testmatrix.Matrix) runner {
+	return runner{Runner: sup.NewRunner(t, m)}
 }
 
 func TestMain(m *testing.M) {
@@ -40,7 +40,9 @@ func TestMain(m *testing.M) {
 		}
 	}
 	exitCode := m.Run()
+	//if sup != nil {
 	sup.PrintSummary()
+	//}
 	os.Exit(exitCode)
 }
